Safari-611.1.10/LayoutTests/fast/images/imageDocument-title-expected.txt

Tests that the title of an ImageDocument is properly set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f.contentDocument.title is "green-24x24.jpg 24×24 pixels"
PASS f.contentDocument.head.getElementsByTagName('title')[0].textContent is "green-24x24.jpg 24×24 pixels"
PASS successfullyParsed is true

TEST COMPLETE
